refactor portfolio transaction routes
