+++ src/app/router_registry.py
@@ -103,6 +103,7 @@ from app.routers.portfolio import router as portfolio_router
 from app.routers.portfolio_activity import router as portfolio_activity_router
 from app.routers.portfolio_book import router as portfolio_book_router
 from app.routers.portfolio_performance import router as portfolio_performance_router
+from app.routers.portfolio_transactions import router as portfolio_transactions_router
 from app.routers.proposal_delivery import router as proposal_delivery_router
 from app.routers.proposal_memo_actions import router as proposal_memo_actions_router
 from app.routers.proposal_memo_evidence import router as proposal_memo_evidence_router
@@ -174,6 +175,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(foundation_router)
     app.include_router(portfolio_router)
     app.include_router(portfolio_activity_router)
+    app.include_router(portfolio_transactions_router)
     app.include_router(portfolio_book_router)
     app.include_router(portfolio_performance_router)
     app.include_router(composite_performance_router)
